+++ bin/commands/users.rb
@@ -31,6 +31,7 @@ command 'users' do |g|
   g.long_desc %( Gets information about a user. )
   g.command 'info' do |c|
     c.flag 'user', desc: 'User to get info on.'
+    c.flag 'include_locale', desc: 'Set this to true to receive the locale for this user. Defaults to false.'
     c.action do |_global_options, options, _args|
       puts JSON.dump($client.users_info(options))
     end
@@ -40,6 +41,7 @@ command 'users' do |g|
   g.long_desc %( Lists all users in a Slack team. )
   g.command 'list' do |c|
     c.flag 'cursor', desc: "Paginate through collections of data by setting the cursor parameter to a next_cursor attribute returned by a previous request's response_metadata. Default value fetches the first 'page' of the collection. See pagination for more detail."
+    c.flag 'include_locale', desc: 'Set this to true to receive the locale for users. Defaults to false.'
     c.flag 'limit', desc: "The maximum number of items to return. Fewer than the requested number of items may be returned, even if the end of the users list hasn't been reached."
     c.flag 'presence', desc: 'Whether to include presence data in the output. Setting this to false improves performance, especially with large teams.'
     c.action do |_global_options, options, _args|