+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.9.2 (Next)
 
 * [#161](https://github.com/slack-ruby/slack-ruby-client/pull/161): Added support for cursor pagination - [@dblock](https://github.com/dblock).
+* [#162](https://github.com/slack-ruby/slack-ruby-client/pull/162): Gracefully close websocket on errno::epipe - [@johanoskarsson](https://github.com/johanoskarsson).
 * Your contribution here.
 
 ### 0.9.1 (8/24/2017)