+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [#163](https://github.com/slack-ruby/slack-ruby-client/pull/164): Use `OpenSSL::X509::DEFAULT_CERT_DIR` and `OpenSSL::X509::DEFAULT_CERT_FILE` for default ca_cert and ca_file - [@leifcr](https://github.com/leifcr).
 * [#161](https://github.com/slack-ruby/slack-ruby-client/pull/161): Added support for cursor pagination - [@dblock](https://github.com/dblock).
+* [#163](https://github.com/slack-ruby/slack-ruby-client/pull/163): Fix UTF8 logging errors in real-time Celluloid component - [@rfwatson](https://github.com/rfwatson).
 * Your contribution here.
 
 ### 0.9.1 (8/24/2017)